+++ .changeset/major-lies-check.md
@@ -1,0 +1,5 @@
+---
+'@keystone-6/core': patch
+---
+
+Removes runtime `node:` imports from `@keystone-6/core` and `@keystone-6/core/context`
